Create OS_Admin__PostInstall_UserHomeDirRestructure.sh
This script will restructure home directory such that all of the following directories,

     Desktop Documents Downloads Music Pictures Public snap Templates Videos

will be replaced by a symbolic link to a corresponding ${username}.${dir} on a non-root partition.  This way, all non-profile related files are stored on non-root partition directly, minimizing impact and effort required to recover root disk if anything happens.  There is still the various .* files in the ~${username} directory which need to be dealt with separately.